fix leaks probleme thanks to abdelah

+++ srcs/init_data.c
@@ -6,7 +6,7 @@
 /*   By: znajdaou <znajdaou@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/31 15:32:00 by znajdaou          #+#    #+#             */
-/*   Updated: 2025/02/05 11:46:43 by znajdaou         ###   ########.fr       */
+/*   Updated: 2025/02/05 13:31:05 by znajdaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ t_data	*ft_init_data(t_data *data, t_fractol selected, char **av)
 		is_anyleft = NULL;
 		data->setx = ft_strtod(av[2], &is_anyleft);
 		data->sety = ft_strtod(av[3], &is_anyleft);
-		if (data->sety > LLONG_MAX || data->setx > LLONG_MAX || is_anyleft)
+		if (data->sety > (double) LLONG_MAX || data->setx > (double)LLONG_MAX || is_anyleft)
 		{
 			ft_free_data(data);
 			ft_exit_error("Error", ERR_INVALID_ARG);
